goose_duck/requirements.txt
# Backend dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# LLM clients (optional)
google-generativeai>=0.3.0
openai>=1.0.0
